streamlit/requirements.txt
streamlit
requests
google-generativeai
faster-whisper
gtts